ccm.c: changed AAD length type for add_auth_data() to 64 bit
To allow AAD with more than 2^32 - 1 bits, the type of the parameter
la has been changed to uint64_t. This also avoids triggering the
type-limits compiler warning.

+++ ccm.c
@@ -73,10 +73,10 @@ block0(size_t M,       /* number of auth bytes */
  * \return     The result is written to \p X.
  */
 static void
-add_auth_data(rijndael_ctx *ctx, const unsigned char *msg, size_t la,
+add_auth_data(rijndael_ctx *ctx, const unsigned char *msg, uint64_t la,
 	      unsigned char B[DTLS_CCM_BLOCKSIZE], 
 	      unsigned char X[DTLS_CCM_BLOCKSIZE]) {
-  size_t i,j; 
+  uint64_t i,j;
 
   rijndael_encrypt(ctx, B, X);
